# gaggiuino-local-profiler/go/internal/system/openapi.yaml
openapi: 3.0.3
info:
  title: GLP REST API
  version: 2.0.0
  description: >
    REST API for the **Gaggiuino Local Profiler** Home Assistant add-on.
    Provides shot storage, coffee library, maintenance tracking, a barista
    orders system, and live machine data.
  contact:
    url: https://github.com/mxkissnr/gaggiuino-local-profiler
servers:
  - url: http://localhost:8099
    description: Direct port (integration, cards)
  - url: /api/hassio_ingress/gaggiuino_local_profiler
    description: HA Ingress (browser — auth via HA session, no token needed)
security:
  - tokenAuth: []
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: X-GLP-Token
      description: >
        64-char hex token, auto-generated on first start, stored in
        `/data/api_token.txt`. Retrieve from `GET /api/status`.
        Requests via HA Ingress bypass this check (HA session is sufficient).
        `GET /api/status` is always public.
    tokenQueryAuth:
      type: apiKey
      in: query
      name: token
      description: >
        Same token as `tokenAuth`, accepted as a `?token=` query parameter
        instead of the `X-GLP-Token` header. Only honored on `GET
        /api/events`, since a browser-native EventSource connection can't
        set custom headers — every other route requires the header.
  schemas:
...
    MqttSettings:
      type: object
      required: [transport, host, port, username, hasPassword, prefix]
      description: >
        GET/POST /api/mqtt/settings response (go/internal/mqtt's SettingsView):
        the redacted view, which reports hasPassword instead of the stored
        password and never echoes the password back (#1050). The POST request
        body is MqttSettingsInput instead, since it additionally carries the
        write-only password/clearPassword pair.
      properties:
        transport: { type: string, enum: [websocket, mqtt] }
        host: { type: string }
        port: { type: integer }
        username: { type: string }
        hasPassword: { type: boolean, description: "Whether a password is stored (never the password itself, #1050)" }
        prefix: { type: string }
    MqttSettingsInput:
      type: object
      required: [transport]
      description: >
        POST /api/mqtt/settings request body (go/internal/mqtt's
        parseSettings). Only transport is required; every other field is
        optional and keeps its current value (or schema default) when omitted.
        password is write-only (absent keeps the stored password unchanged,
        #1050) and clearPassword wipes it regardless (#1062).
      properties:
        transport: { type: string, enum: [websocket, mqtt] }
        host: { type: string }
        port: { type: integer }
        username: { type: string }
        password: { type: string, description: "Write-only: omit to keep the stored password unchanged (#1050); a present value overwrites it unless clearPassword is also set." }
        clearPassword: { type: boolean, description: "Write-only: wipe the stored password regardless of the password field (#1062)." }
        prefix: { type: string }

    # ── MCP ────────────────────────────────────────────────────────────────
...
paths:
...
  /api/mqtt/settings:
    get:
      tags: [Mqtt]
      summary: Current live-data transport settings (Settings-page WebSocket/MQTT toggle + broker connection)
      responses:
        "200":
          description: "{ transport: 'websocket'|'mqtt', host, port, username, hasPassword, prefix } — hasPassword replaces the stored password, which is never echoed back (#1050)"
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MqttSettings' }
    post:
      tags: [Mqtt]
      summary: Save the live-data transport settings. Drops any already-open MQTT session so a changed broker/prefix/credentials takes effect on the next read.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MqttSettingsInput' }
      responses:
        "200": { description: Saved settings, content: { application/json: { schema: { $ref: '#/components/schemas/MqttSettings' } } } }
        "400": { description: Invalid settings payload }
  /api/mqtt/apply-to-machine:
    post:
      tags: [Mqtt]
      summary: One-click apply — points the default machine's own MQTT client (system.mqtt* settings, proxied by #597) at the already-saved broker, so the connection details don't need to be entered twice
      responses:
        "200": { description: Updated machine settings }
        "400": { description: No MQTT broker configured yet }
        "501": { description: machine type does not support the settings proxy }
        "502": { description: Machine unreachable or returned an error }
...